+++ projects/petrknap.cz/config/config.yml
@@ -9,5 +9,7 @@ parameters:
   cache.dir: '%project_dir%/var/cache'
   cache.remote_content.namespace: 'remote_content'
   cache.remote_content.lifetime: 7 * 24 * 3600
+  cache.reveal.namespace: 'reveal'
+  cache.reveal.lifetime: 7 * 24 * 3600
   google_analytics.tracking_id: 'UA-88031264-1'
   backup.dir: '%project_dir%/.backup'
