+++ projects/petrknap.cz/tests/RevealAsPagesTest.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace PetrKnapCz\Test;
+
+use PetrKnapCz\RemoteContent;
+use PetrKnapCz\RemoteContentAccessor;
+use Psr\Cache\CacheItemInterface;
+use Psr\Cache\CacheItemPoolInterface;
+use Symfony\Component\HttpFoundation\Response;
+
+class RevealAsPagesTest extends TestCase
+{
+    public function testModifiesResponse()
+    {
+        $this->markTestIncomplete(); // TODO
+    }
+}
